+++ pretix_mailwizz/urls.py
@@ -1,6 +1,11 @@
 from django.urls import path
 
-from .views import EventSettingsView, OrganizerSettingsView, TestConnectionView
+from .views import (
+    EventSettingsView,
+    OrganizerSettingsView,
+    RetrySyncView,
+    TestConnectionView,
+)
 
 urlpatterns = [
     path(
@@ -17,5 +22,10 @@ urlpatterns = [
         "control/event/<str:organizer>/<str:event>/settings/mailwizz/test/",
         TestConnectionView.as_view(),
         name="event.test",
+    ),
+    path(
+        "control/event/<str:organizer>/<str:event>/settings/mailwizz/retry/",
+        RetrySyncView.as_view(),
+        name="event.retry",
     ),
 ]